requirements.txt
Pillow==8.3.2
moviepy==1.0.3
tkinter==8.6
